+++ frontend/packages/lesson-deep-dive/src/StudentLessonStats/ValidatedLevelsBox.tsx
@@ -2,7 +2,7 @@ import {type CSSProperties, type FC} from 'react';
 
 import RecapStoryCard from '../RecapStoryCard';
 
-import styles from './validated-levels-box.module.scss';
+import styles from './stat-card.module.scss';
 
 interface ValidatedLevelsBoxProps {
   gradient: string;
@@ -43,7 +43,7 @@ const ValidatedLevelsBox: FC<ValidatedLevelsBoxProps> = ({
         className={styles.metricRow}
         style={{'--target-count': validatedLevelsCorrectCount} as CSSProperties}
       >
-        <div className={styles.bigNumber} />
+        <div className={styles.bigNumberAnimated} />
         <p className={styles.caption}>
           of {validatedLevelsTotalCount} validated
           <br />
