--- frontend/packages/lesson-deep-dive/src/StudentLessonStats/levels-attempted-box.module.scss
@@ -1,62 +1,0 @@
-@property --levels-count {
-  syntax: '<integer>';
-  inherits: false;
-  initial-value: 0;
-}
-
-@keyframes count-up {
-  from {
-    --levels-count: 0;
-  }
-
-  to {
-    --levels-count: var(--target-count);
-  }
-}
-
-.body {
-  display: flex;
-  flex-direction: column;
-  gap: 24px;
-}
-
-.metricRow {
-  display: flex;
-  align-items: center;
-  gap: 16px;
-}
-
-.bigNumber {
-  font-family: var(--font-family-main);
-  font-size: clamp(5rem, 18vw, 7.5rem);
-  font-weight: 900;
-  line-height: 1;
-  color: #000;
-  counter-reset: levels-count var(--levels-count);
-  animation: count-up 1.2s cubic-bezier(0.16, 1, 0.3, 1) forwards;
-  flex-shrink: 0;
-
-  &::after {
-    content: counter(levels-count);
-  }
-}
-
-.caption {
-  font-family: var(--font-family-main);
-  font-size: clamp(0.8rem, 2.5vw, 0.95rem);
-  font-weight: 500;
-  color: rgb(0 0 0 / 0.65);
-  line-height: 1.3;
-  margin: 0;
-}
-
-.headline {
-  font-family: var(--font-family-heading);
-  font-size: clamp(1.8rem, 6vw, 2.375rem);
-  font-weight: 900;
-  line-height: 0.95;
-  color: #000;
-  text-transform: uppercase;
-  margin: 0;
-  white-space: pre-line;
-}
