--- frontend/packages/lesson-deep-dive/src/StudentLessonStats/time-spent-box.module.scss
@@ -1,44 +1,0 @@
-.body {
-  display: flex;
-  flex-direction: column;
-  gap: 24px;
-}
-
-.headline {
-  font-family: var(--font-family-heading);
-  font-size: clamp(1.8rem, 6vw, 2.375rem);
-  font-weight: 900;
-  line-height: 0.95;
-  color: #000;
-  text-transform: uppercase;
-  margin: 0;
-  white-space: pre-line;
-}
-
-.headlineRight {
-  text-align: right;
-}
-
-.metricRow {
-  display: flex;
-  align-items: center;
-  gap: 16px;
-}
-
-.bigNumber {
-  font-family: var(--font-family-main);
-  font-size: clamp(5rem, 18vw, 7.5rem);
-  font-weight: 900;
-  line-height: 1;
-  color: #000;
-  flex-shrink: 0;
-}
-
-.caption {
-  font-family: var(--font-family-main);
-  font-size: clamp(0.8rem, 2.5vw, 0.95rem);
-  font-weight: 500;
-  color: rgb(0 0 0 / 0.65);
-  line-height: 1.3;
-  margin: 0;
-}
